+++ src/main.tsx
@@ -7,15 +7,18 @@ import StartupScanProvider from "./context/startup-scan";
 import { ThemeProvider } from "./context/themes";
 import "@/i18n"
 import { RealtimeProvider } from "./context/real-time";
+import { SettingsProvider } from "./context/settings";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
      <React.StrictMode>
-          <RealtimeProvider>
-               <ThemeProvider>
-                    <StartupScanProvider>
-                         <RouterProvider router={router}/>
-                    </StartupScanProvider>
-               </ThemeProvider>
-          </RealtimeProvider>
-     </React.StrictMode>,
+          <SettingsProvider>
+               <RealtimeProvider>
+                    <ThemeProvider>
+                         <StartupScanProvider>
+                              <RouterProvider router={router}/>
+                         </StartupScanProvider>
+                    </ThemeProvider>
+               </RealtimeProvider>
+          </SettingsProvider>
+     </React.StrictMode>
 );